feat:基础功能补充

+++ happy-garden-platform/src/main/java/com/ziyoujiayuan/happygarden/param/RemindSaveParam.java
@@ -18,7 +18,7 @@ public class RemindSaveParam extends DefBaseParam {
 
     private String mode;
 
-    private Byte unit;
+    private Integer unit;
 
     private Date orginTime;
